+++ Cargo.toml
@@ -18,8 +18,6 @@ license-file = "mit-license.md"
 members = ["crates/*"]
 
 [features]
-default = []
-bench = ["criterion", "proptest"]
 dylib = ["wasmer/dylib", "wasmer-engine-dylib", "wasmer-compiler-cranelift"]
 depth_16 = [
     "semaphore-depth-config/depth_16",
@@ -35,10 +33,12 @@ depth_30 = [
 ]
 
 [[bench]]
-name = "criterion"
+name = "lazy_merkle_tree"
 harness = false
-path = "criterion.rs"
-required-features = ["bench", "proptest"]
+
+[[bench]]
+name = "dynamic_merkle_tree"
+harness = false
 
 [dependencies]
 ark-bn254 = { version = "0.3.0" }
@@ -60,12 +60,10 @@ ark-std = { version = "0.3.0", default-features = false, features = [
     "parallel",
 ] }
 color-eyre = "0.6"
-criterion = { version = "0.5", optional = true, features = ["async_tokio"] }
 hex = "0.4.0"
 hex-literal = "0.4"
 num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
 once_cell = "1.8"
-proptest = { version = "1.0", optional = true }
 rand = "0.8.4"
 rayon = "1.5.1"
 itertools = "0.12.0"
@@ -92,6 +90,7 @@ ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features =
 bytemuck = "1.13.1"
 
 [dev-dependencies]
+criterion = { version = "0.5", features = ["async_tokio", "html_reports"] }
 bincode = "1.3.3"
 proptest = "1.0"
 rand_chacha = "0.3.1"
